Random GeoJSON polygon

Random GeoJSON polygon

wkt_polygon(count = 1, num_vertices = 10, max_radial_length = 10,
  bbox = NULL, fmt = 7)

Arguments

count

(integer/numeric) number of Polygons. Default: 1

num_vertices

(integer/numeric) how many coordinates each polygon will contain. Default: 10

max_radial_length

(integer/numeric) maximum distance that a vertex can reach out of the center of the polygon. Units are in degrees latitude (Approximately 69 miles or 111 km). Default: 10

bbox

(integer/numeric) lat/long bounding box for the centers of the polygons, numeric vector of the form west (long), south (lat), east (long), north (lat). optional

fmt

(integer/numeric) number of digits. Default: 7

Value

WKT; a character vector with one or more POLYGON strings

Examples

wkt_polygon()

#> [1] "POLYGON ((-45.1511669 -11.3833549, -47.1802903 -14.1925488, -43.6728289 -15.3265777, -47.6138585 -14.9970712, -44.7482415 -19.4057894, -50.0207655 -15.3106320, -52.7991215 -17.6282695, -49.6559520 -13.8453318, -53.4463459 -11.9087651, -48.9916643 -13.8660146, -45.1511669 -11.3833549))"

wkt_polygon(num_vertices = 3)

#> [1] "POLYGON ((172.4512182 55.8035741, 159.5111389 52.0190870, 166.8133857 60.8695924, 172.4512182 55.8035741))"

wkt_polygon(num_vertices = 4)

#> [1] "POLYGON ((-138.5800881 -14.8841426, -139.7252126 -20.5794128, -144.2603974 -17.7811659, -148.9446096 -10.3412469, -138.5800881 -14.8841426))"

wkt_polygon(num_vertices = 100)

#> [1] "POLYGON ((-108.2773478 -25.2292511, -108.1299953 -24.6538214, -107.3327897 -18.5069666, -107.8997928 -23.4758934, -108.2609895 -26.2419867, -108.3007670 -26.6446850, -108.0538634 -25.7616824, -108.3115768 -26.7803266, -104.6272103 -20.3950452, -103.4399944 -18.8566217, -102.7664475 -18.5492437, -105.5364980 -23.7635413, -106.0940242 -25.1452071, -104.2801984 -23.7127906, -107.9264793 -26.5809912, -108.3415763 -26.8953587, -103.6773370 -24.3155570, -106.0269425 -25.6639008, -105.8366342 -25.7233304, -102.9151332 -24.3831259, -105.8430764 -25.8027746, -101.9361812 -24.4699976, -105.0631620 -25.7951576, -98.5440591 -24.2509052, -98.5523300 -24.8413097, -107.3369683 -26.8379227, -100.0904003 -26.8906949, -101.9782370 -27.2825614, -100.5547871 -27.3834752, -100.5435800 -27.8521562, -107.9249036 -26.9838796, -103.6771351 -28.0617580, -99.5490584 -29.3351956, -102.3014146 -28.7407476, -102.1270298 -28.9694157, -99.3067396 -30.7275206, -101.0930914 -30.2233673, -98.9969602 -31.4821736, -104.5796553 -29.6161429, -105.8833549 -28.7458149, -101.6914467 -31.9748972, -103.3451214 -31.2423062, -107.2264268 -28.1406068, -105.4924862 -30.0320873, -108.3536760 -26.9217709, -105.4263543 -30.1967047, -102.8729915 -32.8867830, -103.3506063 -33.5719967, -106.9160388 -28.9694844, -104.6710380 -33.8007208, -108.2664631 -27.1862413, -107.2980142 -31.1816060, -107.0590036 -32.0897748, -107.6605584 -31.1627368, -108.4530201 -35.0013516, -109.7227292 -36.0617629, -109.6660821 -33.5649278, -108.5891864 -27.8270734, -108.9312487 -28.8188586, -109.9645370 -31.4547144, -113.8391762 -32.2380195, -112.5017055 -30.8897293, -116.1236910 -32.5036355, -110.9630090 -28.7365123, -110.4466864 -28.1916058, -115.3908010 -30.0090205, -111.5143126 -28.3115894, -113.4426364 -28.6946627, -109.6411810 -27.3484696, -115.9119817 -28.6545488, -111.8587250 -27.4665951, -117.4856677 -27.9698789, -114.4657804 -27.5192127, -118.9994244 -27.1488510, -108.7965975 -26.9259259, -115.6311283 -27.0104415, -109.2576831 -26.9255762, -113.2196075 -26.8842381, -116.7234232 -25.5955599, -111.6119674 -26.4480924, -109.5489675 -26.7205515, -109.1658286 -26.7600054, -112.2446154 -26.1063798, -114.5511468 -25.5377647, -115.1986936 -24.0903760, -115.3034970 -23.9965282, -111.7552395 -24.7559988, -109.2453720 -26.3252658, -109.1493022 -26.2100967, -114.0061305 -21.1191856, -109.4824054 -25.7719450, -110.9764615 -23.7461252, -113.9684008 -19.1057377, -112.6153834 -20.6165423, -111.8843007 -21.4214380, -109.9006063 -23.8241513, -112.4173092 -18.0404792, -108.7383894 -25.8420693, -109.6190556 -20.5058621, -108.7383291 -17.0796906, -108.2773478 -25.2292511))"

wkt_polygon(10)

#>  [1] "POLYGON ((44.8782299 7.4841091, 43.3723648 1.8650206, 41.8618376 2.0299869, 42.3672882 0.2250909, 39.0446662 -1.3379778, 36.3997572 5.9870333, 35.0992899 7.4104259, 37.3168742 6.7991822, 36.0541054 8.5711459, 41.1754174 5.1119661, 44.8782299 7.4841091))"                                
#>  [2] "POLYGON ((-54.3547410 32.8226809, -53.0554466 31.8535120, -58.0765545 29.7758254, -59.4570067 28.1847565, -66.1947668 26.1724358, -67.2168929 28.9259361, -64.9859876 29.9351831, -68.8517965 36.5887793, -59.6697729 32.3278760, -59.4907096 38.7922177, -54.3547410 32.8226809))"           
#>  [3] "POLYGON ((77.2167833 16.0389676, 85.4151714 14.0321521, 80.1730012 12.6074797, 70.6551585 8.8101973, 75.1224815 14.4465521, 72.4215733 13.1968900, 73.2866387 14.2771855, 71.2774208 16.1060308, 73.6086262 16.0470873, 74.6559741 14.9962633, 77.2167833 16.0389676))"                       
#>  [4] "POLYGON ((100.1919660 -8.9766803, 98.2993885 -11.4128979, 97.7309314 -13.0401414, 101.8353985 -10.0531432, 101.6346584 -14.5238558, 104.8471441 -15.3794598, 100.3753329 -22.1252004, 96.4447529 -18.6140713, 89.9414471 -22.7616520, 92.9835470 -10.4391703, 100.1919660 -8.9766803))"       
#>  [5] "POLYGON ((-57.5367797 -21.6197728, -55.9251974 -26.2936338, -54.2993848 -31.0762885, -64.8554667 -36.7473928, -63.2416673 -29.8642337, -63.7810236 -30.0230553, -63.6995547 -29.1761992, -63.1652210 -27.5980576, -63.0555509 -26.4684183, -63.1073351 -22.6866056, -57.5367797 -21.6197728))"
#>  [6] "POLYGON ((-46.6113329 -55.3097118, -43.0451095 -60.6924692, -42.3501574 -63.2115715, -46.6165940 -64.7443118, -30.7806221 -63.4081537, -25.8990391 -62.7155755, -54.3260093 -68.1139839, -57.9412175 -67.2231919, -47.9139834 -65.0314570, -54.7005781 -56.3654377, -46.6113329 -55.3097118))"
#>  [7] "POLYGON ((77.1748521 23.5996175, 81.9287289 24.6648302, 74.9783485 20.0682176, 74.7858868 18.8305295, 73.7908059 20.7953604, 72.4072125 20.2723893, 72.1845100 20.1434497, 64.6765776 23.6881076, 71.1940753 22.9373730, 71.2643819 30.8690424, 77.1748521 23.5996175))"                      
#>  [8] "POLYGON ((-61.2641867 -43.2577168, -57.1020111 -44.3319343, -62.9792312 -46.9219611, -62.7899410 -51.8341679, -65.3318026 -42.5791027, -69.6944523 -41.4499144, -70.2406925 -38.3865301, -68.8506301 -38.0681563, -69.1259183 -36.9583440, -67.3304547 -39.4535203, -61.2641867 -43.2577168))"
#>  [9] "POLYGON ((157.3844488 -16.3991578, 157.8673333 -14.4891936, 158.6586560 -15.6096661, 162.7831597 -21.3017162, 155.4551631 -22.6288471, 152.1749839 -26.3401789, 154.6012096 -20.9486299, 152.3169472 -20.2649007, 152.0671495 -19.3554509, 149.4297017 -18.8992738, 157.3844488 -16.3991578))"
#> [10] "POLYGON ((85.0106052 11.0748939, 84.7178416 4.4688738, 93.9853676 7.6604807, 86.8629155 4.9621679, 83.0273424 1.9010750, 80.1885590 -1.2911486, 83.4894429 3.9169963, 83.7423133 4.7126431, 83.0511204 4.9067520, 82.4388798 10.2073341, 85.0106052 11.0748939))"                             

wkt_polygon(bbox = c(50, 50, 60, 60))

#> [1] "POLYGON ((53.3430972 54.1482141, 54.8438410 55.4584848, 49.1110662 49.8240915, 49.8517068 50.8501115, 43.8368758 49.9909770, 44.7292371 50.6565699, 36.3882306 50.2069168, 50.2656617 51.3843396, 48.2778015 52.4977149, 42.3233553 55.8879084, 53.3430972 54.1482141))"