Random GeoJSON polygon

Random GeoJSON polygon

geo_polygon(count = 1, num_vertices = 10, max_radial_length = 10,
  bbox = NULL)

Arguments

count

(integer/numeric) number of Polygons. Default: 1

num_vertices

(integer/numeric) how many coordinates each polygon will contain. Default: 10

max_radial_length

(integer/numeric) maximum distance that a vertex can reach out of the center of the polygon. Units are in degrees latitude (Approximately 69 miles or 111 km). Default: 10

bbox

(integer/numeric) lat/long bounding box for the centers of the polygons, numeric vector of the form west (long), south (lat), east (long), north (lat). optional

Value

GeoJSON; a list with one ore more Polygons in a FeatureCollection

Examples

geo_polygon()

#> $type
#> [1] "FeatureCollection"
#> 
#> $features
#> $features[[1]]
#> $features[[1]]$type
#> [1] "Feature"
#> 
#> $features[[1]]$geometry
#> $features[[1]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[1]]$geometry$coordinates
#> $features[[1]]$geometry$coordinates[[1]]
#> $features[[1]]$geometry$coordinates[[1]][[1]]
#> [1] 173.3263 -28.4603
#> 
#> $features[[1]]$geometry$coordinates[[1]][[2]]
#> [1] 178.36201 -26.53098
#> 
#> $features[[1]]$geometry$coordinates[[1]][[3]]
#> [1] -179.34478  -27.23126
#> 
#> $features[[1]]$geometry$coordinates[[1]][[4]]
#> [1] 173.0557 -35.7697
#> 
#> $features[[1]]$geometry$coordinates[[1]][[5]]
#> [1] 172.71655 -29.08057
#> 
#> $features[[1]]$geometry$coordinates[[1]][[6]]
#> [1] 170.79111 -30.07163
#> 
#> $features[[1]]$geometry$coordinates[[1]][[7]]
#> [1] 169.8455 -30.1258
#> 
#> $features[[1]]$geometry$coordinates[[1]][[8]]
#> [1] 165.88489 -30.41804
#> 
#> $features[[1]]$geometry$coordinates[[1]][[9]]
#> [1] 167.57776 -28.30729
#> 
#> $features[[1]]$geometry$coordinates[[1]][[10]]
#> [1] 171.4842 -24.7779
#> 
#> $features[[1]]$geometry$coordinates[[1]][[11]]
#> [1] 173.3263 -28.4603
#> 
#> 
#> 
#> 
#> $features[[1]]$properties
#> list()
#> 
#> 
#> 

geo_polygon(10)

#> $type
#> [1] "FeatureCollection"
#> 
#> $features
#> $features[[1]]
#> $features[[1]]$type
#> [1] "Feature"
#> 
#> $features[[1]]$geometry
#> $features[[1]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[1]]$geometry$coordinates
#> $features[[1]]$geometry$coordinates[[1]]
#> $features[[1]]$geometry$coordinates[[1]][[1]]
#> [1] 108.22955 -69.67121
#> 
#> $features[[1]]$geometry$coordinates[[1]][[2]]
#> [1] 113.47447 -77.02896
#> 
#> $features[[1]]$geometry$coordinates[[1]][[3]]
#> [1] 120.97765 -78.68021
#> 
#> $features[[1]]$geometry$coordinates[[1]][[4]]
#> [1] 113.15297 -80.18085
#> 
#> $features[[1]]$geometry$coordinates[[1]][[5]]
#> [1] 105.01449 -79.90868
#> 
#> $features[[1]]$geometry$coordinates[[1]][[6]]
#> [1]  83.77685 -83.90544
#> 
#> $features[[1]]$geometry$coordinates[[1]][[7]]
#> [1]  79.36299 -83.08206
#> 
#> $features[[1]]$geometry$coordinates[[1]][[8]]
#> [1]  53.35567 -83.59554
#> 
#> $features[[1]]$geometry$coordinates[[1]][[9]]
#> [1]  88.39603 -77.87475
#> 
#> $features[[1]]$geometry$coordinates[[1]][[10]]
#> [1]  89.62329 -72.04020
#> 
#> $features[[1]]$geometry$coordinates[[1]][[11]]
#> [1] 108.22955 -69.67121
#> 
#> 
#> 
#> 
#> $features[[1]]$properties
#> list()
#> 
#> 
#> $features[[2]]
#> $features[[2]]$type
#> [1] "Feature"
#> 
#> $features[[2]]$geometry
#> $features[[2]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[2]]$geometry$coordinates
#> $features[[2]]$geometry$coordinates[[1]]
#> $features[[2]]$geometry$coordinates[[1]][[1]]
#> [1] 17.92094 30.43838
#> 
#> $features[[2]]$geometry$coordinates[[1]][[2]]
#> [1] 20.11903 27.07772
#> 
#> $features[[2]]$geometry$coordinates[[1]][[3]]
#> [1] 16.36933 26.26266
#> 
#> $features[[2]]$geometry$coordinates[[1]][[4]]
#> [1] 21.03918 25.03709
#> 
#> $features[[2]]$geometry$coordinates[[1]][[5]]
#> [1] 18.53284 23.09629
#> 
#> $features[[2]]$geometry$coordinates[[1]][[6]]
#> [1] 16.08512 26.08395
#> 
#> $features[[2]]$geometry$coordinates[[1]][[7]]
#> [1] 15.34531 22.56944
#> 
#> $features[[2]]$geometry$coordinates[[1]][[8]]
#> [1] 12.49842 21.72339
#> 
#> $features[[2]]$geometry$coordinates[[1]][[9]]
#> [1]  7.910036 22.011349
#> 
#> $features[[2]]$geometry$coordinates[[1]][[10]]
#> [1] 13.92555 29.63934
#> 
#> $features[[2]]$geometry$coordinates[[1]][[11]]
#> [1] 17.92094 30.43838
#> 
#> 
#> 
#> 
#> $features[[2]]$properties
#> list()
#> 
#> 
#> $features[[3]]
#> $features[[3]]$type
#> [1] "Feature"
#> 
#> $features[[3]]$geometry
#> $features[[3]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[3]]$geometry$coordinates
#> $features[[3]]$geometry$coordinates[[1]]
#> $features[[3]]$geometry$coordinates[[1]][[1]]
#> [1] -173.80823   45.37678
#> 
#> $features[[3]]$geometry$coordinates[[1]][[2]]
#> [1] -175.01536   37.44621
#> 
#> $features[[3]]$geometry$coordinates[[1]][[3]]
#> [1] -176.57497   38.23054
#> 
#> $features[[3]]$geometry$coordinates[[1]][[4]]
#> [1] 178.78849  42.17152
#> 
#> $features[[3]]$geometry$coordinates[[1]][[5]]
#> [1] 178.4237  39.1253
#> 
#> $features[[3]]$geometry$coordinates[[1]][[6]]
#> [1] 169.75885  39.73817
#> 
#> $features[[3]]$geometry$coordinates[[1]][[7]]
#> [1] 178.60309  44.44505
#> 
#> $features[[3]]$geometry$coordinates[[1]][[8]]
#> [1] 172.71449  43.20796
#> 
#> $features[[3]]$geometry$coordinates[[1]][[9]]
#> [1] 168.52777  48.25484
#> 
#> $features[[3]]$geometry$coordinates[[1]][[10]]
#> [1] 169.94231  49.97954
#> 
#> $features[[3]]$geometry$coordinates[[1]][[11]]
#> [1] -173.80823   45.37678
#> 
#> 
#> 
#> 
#> $features[[3]]$properties
#> list()
#> 
#> 
#> $features[[4]]
#> $features[[4]]$type
#> [1] "Feature"
#> 
#> $features[[4]]$geometry
#> $features[[4]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[4]]$geometry$coordinates
#> $features[[4]]$geometry$coordinates[[1]]
#> $features[[4]]$geometry$coordinates[[1]][[1]]
#> [1] -166.54284  -43.24133
#> 
#> $features[[4]]$geometry$coordinates[[1]][[2]]
#> [1] -169.41324  -48.19757
#> 
#> $features[[4]]$geometry$coordinates[[1]][[3]]
#> [1] -170.65602  -49.60021
#> 
#> $features[[4]]$geometry$coordinates[[1]][[4]]
#> [1] -163.69975  -50.23085
#> 
#> $features[[4]]$geometry$coordinates[[1]][[5]]
#> [1] -169.93975  -55.59346
#> 
#> $features[[4]]$geometry$coordinates[[1]][[6]]
#> [1] 174.91555 -53.88565
#> 
#> $features[[4]]$geometry$coordinates[[1]][[7]]
#> [1] 171.04438 -49.13218
#> 
#> $features[[4]]$geometry$coordinates[[1]][[8]]
#> [1] 177.58379 -47.22846
#> 
#> $features[[4]]$geometry$coordinates[[1]][[9]]
#> [1] -178.73409  -49.20412
#> 
#> $features[[4]]$geometry$coordinates[[1]][[10]]
#> [1] 178.59232 -42.56677
#> 
#> $features[[4]]$geometry$coordinates[[1]][[11]]
#> [1] -166.54284  -43.24133
#> 
#> 
#> 
#> 
#> $features[[4]]$properties
#> list()
#> 
#> 
#> $features[[5]]
#> $features[[5]]$type
#> [1] "Feature"
#> 
#> $features[[5]]$geometry
#> $features[[5]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[5]]$geometry$coordinates
#> $features[[5]]$geometry$coordinates[[1]]
#> $features[[5]]$geometry$coordinates[[1]][[1]]
#> [1] 25.639947 -7.396703
#> 
#> $features[[5]]$geometry$coordinates[[1]][[2]]
#> [1] 21.285829 -8.400151
#> 
#> $features[[5]]$geometry$coordinates[[1]][[3]]
#> [1] 19.305946 -8.829688
#> 
#> $features[[5]]$geometry$coordinates[[1]][[4]]
#> [1]  26.96087 -13.28881
#> 
#> $features[[5]]$geometry$coordinates[[1]][[5]]
#> [1] 19.050567 -9.546623
#> 
#> $features[[5]]$geometry$coordinates[[1]][[6]]
#> [1]  13.09127 -17.10599
#> 
#> $features[[5]]$geometry$coordinates[[1]][[7]]
#> [1]  12.93642 -10.71384
#> 
#> $features[[5]]$geometry$coordinates[[1]][[8]]
#> [1] 17.125030 -9.100276
#> 
#> $features[[5]]$geometry$coordinates[[1]][[9]]
#> [1] 10.574597 -4.973843
#> 
#> $features[[5]]$geometry$coordinates[[1]][[10]]
#> [1] 18.645620 -8.554908
#> 
#> $features[[5]]$geometry$coordinates[[1]][[11]]
#> [1] 25.639947 -7.396703
#> 
#> 
#> 
#> 
#> $features[[5]]$properties
#> list()
#> 
#> 
#> $features[[6]]
#> $features[[6]]$type
#> [1] "Feature"
#> 
#> $features[[6]]$geometry
#> $features[[6]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[6]]$geometry$coordinates
#> $features[[6]]$geometry$coordinates[[1]]
#> $features[[6]]$geometry$coordinates[[1]][[1]]
#> [1] 24.76554 34.34676
#> 
#> $features[[6]]$geometry$coordinates[[1]][[2]]
#> [1] 35.18519 29.95280
#> 
#> $features[[6]]$geometry$coordinates[[1]][[3]]
#> [1] 35.13898 26.76900
#> 
#> $features[[6]]$geometry$coordinates[[1]][[4]]
#> [1] 29.45087 22.76209
#> 
#> $features[[6]]$geometry$coordinates[[1]][[5]]
#> [1] 28.26715 21.30334
#> 
#> $features[[6]]$geometry$coordinates[[1]][[6]]
#> [1] 18.94752 26.71810
#> 
#> $features[[6]]$geometry$coordinates[[1]][[7]]
#> [1] 15.97795 28.06442
#> 
#> $features[[6]]$geometry$coordinates[[1]][[8]]
#> [1] 13.18269 29.25527
#> 
#> $features[[6]]$geometry$coordinates[[1]][[9]]
#> [1] 18.21163 34.31100
#> 
#> $features[[6]]$geometry$coordinates[[1]][[10]]
#> [1] 23.72905 31.72115
#> 
#> $features[[6]]$geometry$coordinates[[1]][[11]]
#> [1] 24.76554 34.34676
#> 
#> 
#> 
#> 
#> $features[[6]]$properties
#> list()
#> 
#> 
#> $features[[7]]
#> $features[[7]]$type
#> [1] "Feature"
#> 
#> $features[[7]]$geometry
#> $features[[7]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[7]]$geometry$coordinates
#> $features[[7]]$geometry$coordinates[[1]]
#> $features[[7]]$geometry$coordinates[[1]][[1]]
#> [1] -49.28983 -23.01288
#> 
#> $features[[7]]$geometry$coordinates[[1]][[2]]
#> [1] -52.67528 -28.80544
#> 
#> $features[[7]]$geometry$coordinates[[1]][[3]]
#> [1] -45.68675 -33.73057
#> 
#> $features[[7]]$geometry$coordinates[[1]][[4]]
#> [1] -52.96158 -29.00736
#> 
#> $features[[7]]$geometry$coordinates[[1]][[5]]
#> [1] -63.02404 -26.89496
#> 
#> $features[[7]]$geometry$coordinates[[1]][[6]]
#> [1] -54.61238 -27.93163
#> 
#> $features[[7]]$geometry$coordinates[[1]][[7]]
#> [1] -57.00838 -24.70216
#> 
#> $features[[7]]$geometry$coordinates[[1]][[8]]
#> [1] -55.08604 -26.37399
#> 
#> $features[[7]]$geometry$coordinates[[1]][[9]]
#> [1] -55.70643 -21.85343
#> 
#> $features[[7]]$geometry$coordinates[[1]][[10]]
#> [1] -53.55443 -26.80018
#> 
#> $features[[7]]$geometry$coordinates[[1]][[11]]
#> [1] -49.28983 -23.01288
#> 
#> 
#> 
#> 
#> $features[[7]]$properties
#> list()
#> 
#> 
#> $features[[8]]
#> $features[[8]]$type
#> [1] "Feature"
#> 
#> $features[[8]]$geometry
#> $features[[8]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[8]]$geometry$coordinates
#> $features[[8]]$geometry$coordinates[[1]]
#> $features[[8]]$geometry$coordinates[[1]][[1]]
#> [1] -50.73780 -62.01922
#> 
#> $features[[8]]$geometry$coordinates[[1]][[2]]
#> [1] -53.55285 -63.53644
#> 
#> $features[[8]]$geometry$coordinates[[1]][[3]]
#> [1] -51.09035 -63.83932
#> 
#> $features[[8]]$geometry$coordinates[[1]][[4]]
#> [1] -41.51569 -70.44475
#> 
#> $features[[8]]$geometry$coordinates[[1]][[5]]
#> [1] -56.97516 -65.23826
#> 
#> $features[[8]]$geometry$coordinates[[1]][[6]]
#> [1] -58.92359 -63.26426
#> 
#> $features[[8]]$geometry$coordinates[[1]][[7]]
#> [1] -55.13298 -63.42728
#> 
#> $features[[8]]$geometry$coordinates[[1]][[8]]
#> [1] -69.49725 -57.64435
#> 
#> $features[[8]]$geometry$coordinates[[1]][[9]]
#> [1] -57.89588 -61.67707
#> 
#> $features[[8]]$geometry$coordinates[[1]][[10]]
#> [1] -55.67972 -55.68358
#> 
#> $features[[8]]$geometry$coordinates[[1]][[11]]
#> [1] -50.73780 -62.01922
#> 
#> 
#> 
#> 
#> $features[[8]]$properties
#> list()
#> 
#> 
#> $features[[9]]
#> $features[[9]]$type
#> [1] "Feature"
#> 
#> $features[[9]]$geometry
#> $features[[9]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[9]]$geometry$coordinates
#> $features[[9]]$geometry$coordinates[[1]]
#> $features[[9]]$geometry$coordinates[[1]][[1]]
#> [1] -150.23979  -12.19101
#> 
#> $features[[9]]$geometry$coordinates[[1]][[2]]
#> [1] -148.98358  -11.87234
#> 
#> $features[[9]]$geometry$coordinates[[1]][[3]]
#> [1] -148.35383  -20.43397
#> 
#> $features[[9]]$geometry$coordinates[[1]][[4]]
#> [1] -144.88169  -26.24096
#> 
#> $features[[9]]$geometry$coordinates[[1]][[5]]
#> [1] -151.77047  -25.28335
#> 
#> $features[[9]]$geometry$coordinates[[1]][[6]]
#> [1] -156.47670  -25.19971
#> 
#> $features[[9]]$geometry$coordinates[[1]][[7]]
#> [1] -158.17133  -25.32737
#> 
#> $features[[9]]$geometry$coordinates[[1]][[8]]
#> [1] -152.94725  -19.17789
#> 
#> $features[[9]]$geometry$coordinates[[1]][[9]]
#> [1] -156.21243  -13.92668
#> 
#> $features[[9]]$geometry$coordinates[[1]][[10]]
#> [1] -151.41702  -14.31432
#> 
#> $features[[9]]$geometry$coordinates[[1]][[11]]
#> [1] -150.23979  -12.19101
#> 
#> 
#> 
#> 
#> $features[[9]]$properties
#> list()
#> 
#> 
#> $features[[10]]
#> $features[[10]]$type
#> [1] "Feature"
#> 
#> $features[[10]]$geometry
#> $features[[10]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[10]]$geometry$coordinates
#> $features[[10]]$geometry$coordinates[[1]]
#> $features[[10]]$geometry$coordinates[[1]][[1]]
#> [1] 152.61101 -14.51637
#> 
#> $features[[10]]$geometry$coordinates[[1]][[2]]
#> [1] 156.05262 -15.44952
#> 
#> $features[[10]]$geometry$coordinates[[1]][[3]]
#> [1] 156.1326 -19.6731
#> 
#> $features[[10]]$geometry$coordinates[[1]][[4]]
#> [1] 152.9374 -21.5547
#> 
#> $features[[10]]$geometry$coordinates[[1]][[5]]
#> [1] 149.33647 -25.83087
#> 
#> $features[[10]]$geometry$coordinates[[1]][[6]]
#> [1] 149.15209 -24.84025
#> 
#> $features[[10]]$geometry$coordinates[[1]][[7]]
#> [1] 146.72431 -27.26147
#> 
#> $features[[10]]$geometry$coordinates[[1]][[8]]
#> [1] 143.68725 -20.39584
#> 
#> $features[[10]]$geometry$coordinates[[1]][[9]]
#> [1] 146.49357 -19.55122
#> 
#> $features[[10]]$geometry$coordinates[[1]][[10]]
#> [1] 146.78286 -11.93651
#> 
#> $features[[10]]$geometry$coordinates[[1]][[11]]
#> [1] 152.61101 -14.51637
#> 
#> 
#> 
#> 
#> $features[[10]]$properties
#> list()
#> 
#> 
#> 

geo_polygon(bbox = c(50, 50, 60, 60))

#> $type
#> [1] "FeatureCollection"
#> 
#> $features
#> $features[[1]]
#> $features[[1]]$type
#> [1] "Feature"
#> 
#> $features[[1]]$geometry
#> $features[[1]]$geometry$type
#> [1] "Polygon"
#> 
#> $features[[1]]$geometry$coordinates
#> $features[[1]]$geometry$coordinates[[1]]
#> $features[[1]]$geometry$coordinates[[1]][[1]]
#> [1] 51.54547 55.28739
#> 
#> $features[[1]]$geometry$coordinates[[1]][[2]]
#> [1] 62.16176 56.09665
#> 
#> $features[[1]]$geometry$coordinates[[1]][[3]]
#> [1] 62.29150 55.00115
#> 
#> $features[[1]]$geometry$coordinates[[1]][[4]]
#> [1] 56.20678 44.77152
#> 
#> $features[[1]]$geometry$coordinates[[1]][[5]]
#> [1] 50.42861 51.66230
#> 
#> $features[[1]]$geometry$coordinates[[1]][[6]]
#> [1] 46.59109 44.34078
#> 
#> $features[[1]]$geometry$coordinates[[1]][[7]]
#> [1] 44.42743 53.74798
#> 
#> $features[[1]]$geometry$coordinates[[1]][[8]]
#> [1] 39.95936 54.25047
#> 
#> $features[[1]]$geometry$coordinates[[1]][[9]]
#> [1] 40.42974 59.32499
#> 
#> $features[[1]]$geometry$coordinates[[1]][[10]]
#> [1] 50.72923 54.29451
#> 
#> $features[[1]]$geometry$coordinates[[1]][[11]]
#> [1] 51.54547 55.28739
#> 
#> 
#> 
#> 
#> $features[[1]]$properties
#> list()
#> 
#> 
#>